Add base class fo unique and graduated layers-styles
