+++ addon/templates/components/layers-styles-editor.hbs
@@ -1,37 +1,35 @@
-  {{#unless _leafletLayerLoadingIsError}}
-    {{#unless _leafletLayerIsLoading}}
-      <div class="field">
-        <label>
-          {{t "components.layers-styles-editor.layer-style-dropdown.caption"}}
-        </label>
-        {{flexberry-dropdown
-          class="fluid"
-          items=_availableLayerStylesCaptions
-          value=_selectedLayerStyleCaption
-        }}
-      </div>
-      {{#if (known-for-type "component" (concat "layers-styles/" styleSettings.type))}}
-        {{component (concat "layers-styles/" styleSettings.type)
-          styleSettings=styleSettings
-          displaySettings=displaySettings
-          layerType=layerType
-          leafletLayer = _leafletLayer
-        }}
-      {{else}}
-        {{ui-message
-          type="error"
-          message=(t
-            "components.layers-styles-editor.layer-style-component-not-implemented.message"
-            componentName=(concat "layers-styles/" styleSettings.type)
-          )
-        }}
-      {{/if}}
-    {{else}}
-      <div class="ui form loading" style="height: 100%;">
-      </div>
-    {{/unless}}
+<div class={{concat "ui inverted dimmer" (if _leafletLayerIsLoading " active" " disabled") }}>
+  <div class="ui loader"></div>
+</div>
+
+{{#unless _leafletLayerLoadingIsError}}
+  <div class="field">
+    <label>
+      {{t "components.layers-styles-editor.layer-style-dropdown.caption"}}
+    </label>
+    {{flexberry-dropdown
+      class="fluid"
+      items=_availableLayerStylesCaptions
+      value=_selectedLayerStyleCaption
+    }}
+  </div>
+  {{#if (known-for-type "component" (concat "layers-styles/" styleSettings.type))}}
+    {{component (concat "layers-styles/" styleSettings.type)
+      styleSettings=styleSettings
+      displaySettings=displaySettings
+      layerType=layerType
+      leafletLayer = _leafletLayer
+    }}
   {{else}}
-    <div class="field">
-      {{t "components.layers-styles-editor.layer-style-layer-loading-error.caption"}}
-    </div>
-  {{/unless}}
+    {{ui-message
+      type="error"
+      message=(t "components.layers-styles-editor.layer-style-component-not-implemented.message"
+        componentName=(concat "layers-styles/" styleSettings.type)
+      )
+    }}
+  {{/if}}
+{{else}}
+  <div class="field">
+    {{t "components.layers-styles-editor.layer-style-layer-loading-error.caption"}}
+  </div>
+{{/unless}}
