+++ app/components/layers-styles/categorized/base-custom-layer-style.js
@@ -1,0 +1,1 @@
+export { default } from 'ember-flexberry-gis/components/layers-styles/categorized/base-custom-layer-style';
